feat(subdomains): add list subdomains functionality
Adds a new API endpoint to list all subdomains with pagination.

This endpoint allows users to retrieve a list of all subdomains in the system, paginated by 15 results per page. The endpoint also provides the total number of subdomains in the system, as well as the current page and total number of pages.